requirements.txt.txt
matplotlib==3.7.0
numpy==1.23.5
pillow==9.4.0
streamlit==1.28.0
ultralytics==8.1.20